requirements.txt
asn1crypto==0.24.0
astropy==3.1.2
certifi==2019.3.9
cffi==1.12.3
cryptography==2.6.1
future==0.17.1
numpy==1.16.3
pycparser==2.19
python-telegram-bot==12.0.0b1
pytz==2019.1
schedule==0.6.0
six==1.12.0
tornado==6.0.2
